+++ auth-db/queries/email-auth/index.ts
@@ -1,2 +1,3 @@
 export * from "./create.ts";
 export * from "./get-by-email.ts";
+export * from "./update-verification.ts";
